+++ jobbergate-agent/jobbergate_agent/tasks.py
@@ -8,6 +8,7 @@ from loguru import logger
 
 from jobbergate_agent.clients.cluster_api import backend_client
 from jobbergate_agent.internals.update import self_update_agent
+from jobbergate_agent.jobbergate.report_health import report_health_status
 from jobbergate_agent.jobbergate.submit import submit_pending_jobs
 from jobbergate_agent.jobbergate.update import update_active_jobs
 from jobbergate_agent.settings import SETTINGS
@@ -36,6 +37,16 @@ def pending_submissions_task(scheduler: BaseScheduler) -> Job:
     Schedule a task to submit pending jobs every ``TASK_JOBS_INTERVAL_SECONDS`` seconds.
     """
     return scheduler.add_job(submit_pending_jobs, "interval", seconds=SETTINGS.TASK_JOBS_INTERVAL_SECONDS)
+
+
+def status_report_task(scheduler: BaseScheduler) -> Job:
+    """
+    Schedule a task to report the status.
+    """
+    seconds_between_calls = SETTINGS.TASK_JOBS_INTERVAL_SECONDS
+    return scheduler.add_job(
+        report_health_status, "interval", seconds=seconds_between_calls, kwargs={"interval": seconds_between_calls}
+    )
 
 
 @logger_wraps()